+++ src/services/post/noticeMutations.ts
@@ -5,16 +5,11 @@ import { axiosInstance } from '@/services/axiosInstance';
 import { AxiosError } from 'axios';
 
 export const usePostNoticeSessions = () => {
-  return useMutation<
-    boolean,
-    AxiosError,
-    { phoneNumber: string; sessionId: number }
-  >({
+  return useMutation<boolean, AxiosError, { sessionId: number }>({
     mutationFn: async requestBody => {
-      const { phoneNumber, sessionId } = requestBody;
+      const { sessionId } = requestBody;
       const { data } = await axiosInstance.post(
         `/notices/sessions/${sessionId}/application`,
-        { phoneNumber },
       );
       return data;
     },
@@ -22,17 +17,10 @@ export const usePostNoticeSessions = () => {
 };
 
 export const usePostSessionApply = () => {
-  return useMutation<
-    boolean,
-    AxiosError,
-    { sessionId: number; phoneNumber: string }
-  >({
-    mutationFn: async ({ sessionId, phoneNumber }) => {
+  return useMutation<boolean, AxiosError, { sessionId: number }>({
+    mutationFn: async ({ sessionId }) => {
       const { data } = await axiosInstance.post(
         `/notices/sessions/${sessionId}/application`,
-        {
-          body: { phoneNumber },
-        },
       );
       return data;
     },
