+++ src/services/post/noticeQueries.ts
@@ -4,7 +4,6 @@ import { useInfiniteQuery, useQuery } from '@tanstack/react-query';
 
 import { axiosInstance } from '@/services/axiosInstance';
 
-import { SessionStatus } from '@/constants/serviceConstant';
 import { NoticeDto, NoticeFilter } from '@/types';
 import { isInThisWeek } from '@/utils';
 
@@ -80,36 +79,6 @@ export const useGetThisWeekNoticeList = () => {
   return useQuery({
     queryKey: ['useGetThisWeekNoticeList'],
     queryFn: getThisWeekNotice,
-    retry: false,
-  });
-};
-
-export const useGetSessionStatus = ({
-  sessionId,
-  page = 1,
-  size = 10,
-  searchParticipantStatus = 'WAIT',
-}: {
-  sessionId: number;
-  page?: number;
-  size?: number;
-  searchParticipantStatus?: SessionStatus;
-}) => {
-  const getSessionStatus = async () => {
-    const { data } = await axiosInstance.get(`/notices/sessions/${sessionId}`, {
-      params: {
-        sessionId,
-        page,
-        size,
-        searchParticipantStatus,
-      },
-    });
-    return data;
-  };
-  return useQuery({
-    queryKey: ['useGetSessionStatus', sessionId],
-    queryFn: getSessionStatus,
-    enabled: !!Number(sessionId),
     retry: false,
   });
 };
